+++ CHANGELOG.md
@@ -23,3 +23,8 @@ July 23, 2018
 
 - version "1.0.24"
 - spatial_tree_t::find_cell
+
+September 21, 2018
+
+- version "1.0.29"
+- database::make_shared_table